requirements.txt
neo4j==5.15.0
pandas==2.1.4
python-dotenv==1.0.0
openai==1.12.0
flask==3.0.0
requests==2.31.0
google-genai
dotenv
neo4j